requirements.txt
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==4.0
django-allauth==0.47.0
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
virtualenv==20.10.0